requirements_test.txt
coverage==4.3.4
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
django-storages==1.5.2
boto==2.46.1
